// src/subapps/projects/fusionContext.d.ts
export declare const WORKFLOW_STEP_CONTEXT: {
  '@context': {
    '@base': string;
    '@vocab': string;
    nxv: string;
    input: string;
    activity: string;
    name: {
      '@id': string;
    };
    description: {
      '@id': string;
    };
  }[];
  '@id': string;
};
export declare const FUSION_TABLE_CONTEXT: {
  '@context': {
    '@base': string;
    '@vocab': string;
    name: {
      '@id': string;
    };
    description: {
      '@id': string;
    };
  }[];
  '@id': string;
};
export declare const PROJECT_METADATA_CONTEXT: {
  '@context': {
    '@base': string;
    '@vocab': string;
    name: {
      '@id': string;
    };
    description: {
      '@id': string;
    };
  }[];
  '@id': string;
};